requirements.txt
cuda==11.6.1
datasets==2.15.0
matplotlib==3.6.2
numpy==1.23.5
pandas==1.5.2
Pillow==9.3.0
python==3.10.9
scikit-learn==1.2.1
scipy==1.10.0
seaborn==0.12.2
torch==1.13.1
torchvision==0.14.1
tqdm==4.65.0
